+++ src/domain/interfaces/auth-api.interface.ts
@@ -6,7 +6,6 @@ export interface IAuthApi {
   generateNonce(): string;
 
   verifyMessage(args: {
-    address: `0x${string}`;
     message: SiweMessage;
     signature: `0x${string}`;
   }): Promise<boolean>;
